Steps_1_to_4/requirements.txt
pandas==2.0.1
python-dotenv==1.0.1
pytz==2022.1
Requests==2.31.0
setuptools==59.6.0